Use last match of discord in file path
Change filepath matching so that if discord appears multiple times in a
traceback line, the last one is used instead of the first one as the
base for finding paths.  Fixes nothing matching when the lib is in a
directory with discord in the name, which is the case in for example
venvs.